requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy==1.20.2
pandas==1.2.4
protobuf==3.15.8
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
tabulate==0.8.9
termcolor==1.1.0
urllib3==1.26.4
yfinance==0.1.59
